Merge with main.
At least it compiles and things seem to work at least on the surface, apart from newsfeedcontroller that I couldn't get to work. Some warnings like this, the id seems to change from time to time:

Spawned NetworkObject was expected to exist but does not for Id 6. This may occur if you sent a NetworkObject reference which does not exist, be it destroyed or if the client does not have visibility.